--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -60,6 +60,7 @@ end
 @safetestset "Soil CO2 parameterization tests" begin
     include("standalone/Soil/Biogeochemistry/co2_parameterizations.jl")
 end
+
 @safetestset "Soil climate drivers tests" begin
     include("standalone/Soil/climate_drivers.jl")
 end