--- a/test/standalone/Soil/climate_drivers.jl
+++ b/test/standalone/Soil/climate_drivers.jl
@@ -142,6 +142,8 @@ for FT in (Float32, Float64)
                 :θ_l,
                 :T,
                 :κ,
+                :bidiag_matrix_scratch,
+                :full_bidiag_matrix_scratch,
                 :turbulent_fluxes,
                 :R_n,
                 :top_bc,