<!--- THESE LINES ARE COMMENTED -->
## Purpose 
This PR computes intermediate quantities in the soil Jacobian for the Richards and EnergyHydrology model. This circumvents performance issues in ClimaCore when many matrix field broadcast operations are fused.

This improves the Richards benchmark time from 1.5s to 0.6s, and the snowy land benchmark time from 3.29s to 0.741s. It changes the long run SYPD as:
snowy land model: 20 SYPD -> ~60 SYPD


## To-do
- understand why more NaNs appear in the full soil long run (this is due to PR #1113; not this current PR)

## Content
- adds bidiagonal matrix row fields to scratch space for the soil model, uses these in the jacobian computation
- does the same with a covariant3vector field
- updates benchmark times


<!---
Review checklist

I have:
- followed the codebase contribution guide: https://clima.github.io/ClimateMachine.jl/latest/Contributing/
- followed the style guide: https://clima.github.io/ClimateMachine.jl/latest/DevDocs/CodeStyle/
- followed the documentation policy: https://github.com/CliMA/policies/wiki/Documentation-Policy
- checked that this PR does not duplicate an open PR.

In the Content, I have included 
- relevant unit tests, and integration tests, 
- appropriate docstrings on all functions, structs, and modules, and included relevant documentation.

-->

----
- [ ] I have read and checked the items on the review checklist.