@juliasloan25 2025-04-25
Just a note: this PR is a breaking change because it introduces a `subsurface_face` space in the domain, which needs to be allocated when the coupler sets up the land domain. So our next release should be minor (0.16.0).

This wouldn't be a problem if ClimaCoupler used the `global_domain` function, which we'll switch to soon :)